Changes to fix the Windows dll linkage problems (undefined/duplicate symbol problems). Removed all implementations from header files where suitable (i.e. non-template). Applied a fix for handling the time for License in 32-bit environments after year 2038.

+++ hazelcast/src/hazelcast/client/serialization/pimpl/Data.cpp
@@ -36,7 +36,7 @@ namespace hazelcast {
                     if (data.get() != 0 && data->size() > 0 && data->size() < DATA_OFFSET) {
                         char msg[100];
                         sprintf(msg, "Provided buffer should be either empty or "
-                                "should contain more than %d bytes! Provided buffer size:%ld", DATA_OFFSET, data->size());
+                                "should contain more than %u bytes! Provided buffer size:%lu", DATA_OFFSET, (unsigned long)data->size());
                         throw exception::IllegalArgumentException("Data::setBuffer", msg);
                     }
                 }
